backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
groq>=0.4.0
huggingface_hub>=0.23.0
httpx
